refactor: Move global.css to src/styles/
