+++ src/pages/_app.js
@@ -1,4 +1,4 @@
-import '../app/globals.css'
+import '../styles/globals.css'
 
 export default function App({ Component, pageProps }) {
     return <Component {...pageProps} />
